app/requirements.txt
streamlit==0.52.1
numpy==1.16.6
mtcnn==0.1.0
Pillow==7.2.0
tensorflow-cpu==2.2.0
setuptools==41.0.0
opencv-python-headless==4.3.0.36